+++ decidim-initiatives/app/assets/config/decidim_initiatives_manifest.js
@@ -1,7 +1,5 @@
 //= link decidim/initiatives/social_share.js
 //= link decidim/initiatives/application.js
-//= link decidim/initiatives/application.css
-//= link decidim/initiatives/social_share.css
 //= link decidim/initiatives/identity_selector_dialog.js
 //= link decidim/initiatives/scoped_type.js
 //= link decidim/initiatives/admin/invite_users.js